Moved querying for category matches to journal model using query object
Query object is populated with the corresponding category matches and it is parsed in the journal model to finally
apply the matches with gl_journal_set_matches().After applying the matches, journal is repopulated accordingly.
The gl_journal_set_matches() function was altered to accept a GPtrArray for the matches instead of const array
with constant string as we get the category matches in a GPtrArray while parsing in journal model.