Moved search matches to journal model using query API
The querying for user entered search text is now done by parsing the query object in journal model and it is
populated accordingly. The query object is populated in the view. Earlier, the searching work was done by
listbox_search_filter_func() but now it is transferred to the query API. A new enum 'GlSearchQueryType' was
introduced to differentiate between exact and substring matches.The search related functions from eventviewlist
are also removed.