Implemented Search Popover with Parameter filtering
The search popover is implemented in gl-eventviewlist including it's event handlers. We use GtkTreeView to
display the parameters in the popover.It's model is defined in the gl-searchpopover GtkBuilder file.Now, the
entries in journal can be filtered by selecting the appropriate option in  the search popover. Also, the
gl-journal-model and gl-journal classes were modified to handle the new journal parameters which are shown
in the search popover.The tokenize_search_string() function in gl-journal-model was changed to handle
slashes for Executable Path parameter.