Added search type option in search popover
Added a Radio button group to select between "Exact" and "Substring" search type. We hide the "Search
Type" option through a GtkRevealer if "All Available Fields" is selected in the parameter filtering options
as exact search doesn't  make sense in that case.The GlJournalModel class was also modified to handle the
case when all search fields are exact and there are no substring matches. Also, it was modified to handle
the case when the exact search match string has a null or empty field value in which case we do not pass it
to gl_journal_set_matches().