Implemented Search Popover with Parameter filtering
The search popover is implemented as a seperate class called GlSearchPopover. We use GtkTreeView to
display the parameters in the popover.It's model is defined in the gl-searchpopover GtkBuilder file.Now, the
entries in journal can be filtered by selecting the appropriate option in  the search popover. Also, the
gl-journal-model and gl-journal classes were modified to handle the new journal parameters which are shown
in the search popover.The GScannerConfig in tokenize_search_string() function in gl-journal-model was
modified to detect characters required in values of certain journal fields like _KERNEL_DEVICE,
_SYSTEMD_UNIT, _EXE.