Add search type option to search popover
A radio button group was addded to select between "Exact"
and "Substring" search types. We hide the search type radio
button group if "All Available Fields" is selected as exact
search doesn't make sense in that case. The GlJournalModel class
was modified to handle the case when all search fields are exact
and there are no substring matches. Also, it was modified to handle
the case when the exact search match string has a null or empty
value in which case we do not pass it to gl_journal_set_matches().
Also, a GL_QUERY_* prefix was added to the enum nicks for making
it consistent with the enum name GlQuerySearchType.

https://bugzilla.gnome.org/show_bug.cgi?id=767996