Add journal timestamp range filters
The journal timestamp range filters in search popover can be
used to set the timestamp range of journal model.The default
journal timestamp range is the currently selected boot from
the boot selection menu in toolbar.

The query object was modified to include two
new parameters for storing the start and end timestamps
of the journal.

The gl_journal_set_matches() function was modified to
remove setting the boot match if no boot match is specified.

The search popover range button label is made consistent when
user selects boot through the toolbar boot selection menu.

https://bugzilla.gnome.org/show_bug.cgi?id=767996