Add range filtering to search popover
The range filtering option in search popover can used to
set the timestamp range of journal.The default journal range
is current boot. The query object was modified to include two
new parameters for storing the start and end timestamps
of the journal. The gl-journal-set-matches() was modified to
remove setting the boot match if no boot match is specified.
The search popover range button label is made consistent when
user selects boot through the toolbar boot selection menu.

https://bugzilla.gnome.org/show_bug.cgi?id=767996