Implement handle-launch-search method
We implement the handle-launch-search method
of the shell search provider interface in this
commit.

When the user clicks on the GNOME Logs icon
besides the search provider results, a new
window will open showing all the results for the
user entered search term in GNOME Shell Overview
search bar.

The search results in the window are shown from
"All" category so that user can see all possible
results for his search term.

https://bugzilla.gnome.org/show_bug.cgi?id=753470