Define GlRowEntry object and it's API
The GlRowEntry object includes flags/variables to store information
about the compression of journal entries and pass it on from the
model to view.

There are three types of row entries: uncompressed, compressed, header.

If a row entry is a header, then it also stores the compressed
journal entries which are represented by the header.