Add initial implementation of shell search provider
The following three methods of the shell search provider interface
are implemented:

* handle-get-initial-result-set
* handle-get-subsearch-result-set
* handle-get-result-metas

The user entered search text in gnome shell overview
is searched in all the available journal fields for each
journal entry.

In the search provider results we show the journal entry message
as the title of each result. Process name and message time are shown
in the description of the result.

In each search invocation, a maximum of 5 results are fetched and
the logs are shown in descending order of timestamp.

Current boot is used as the journal timestamp range.

https://bugzilla.gnome.org/show_bug.cgi?id=753470