Implement handle-launch-search method
The "handle-launch-search" method of the shell
search provider interface is implemented in this
commit.

When the user clicks on the GNOME Logs icon
besides the search provider results, existing
window is reused or else a new window opens showing
all the results for the user entered search term in
GNOME Shell Overview.

The search results in the window are shown from
"All" category, by default, so that user can see
all possible results for his search term.

https://bugzilla.gnome.org/show_bug.cgi?id=753470