Implement handle-activate-result method
The "handle-activate-result" shell search provider
interface method is implemented in this commit.

In the GNOME Shell overview search provider results,
if a search result is clicked , a new window
opens which shows the detailed information about the
result in a GtkPopover.

In the newly opened window, the event toolbar and
category list is hidden.

https://bugzilla.gnome.org/show_bug.cgi?id=753470